clean code
# message auto-generated for no-merge-commit merge:
merge yuwangang into master

clean code

Created-by: msprofDeveloper

Match-id-21d1fe398742c090bc591e5c44147090054a3197

+++ analysis/csrc/application/summary/task_time_assembler.cpp
@@ -96,7 +96,7 @@ void TaskTimeAssembler::AssembleTaskTime(const std::vector<AscendTaskData> &asce
         auto row = {std::to_string(ascendTaskDatum.deviceId),
                     opName, taskType, std::to_string(taskId.streamId),
                     std::to_string(taskId.taskId),
-                    DivideByPowersOfTenWithPrecision(ascendTaskDatum.duration),
+                    DivideByPowersOfTenWithPrecision(static_cast<uint64_t>(ascendTaskDatum.duration)),
                     DivideByPowersOfTenWithPrecision(ascendTaskDatum.timestamp) + DIVIDE_CHAR,
                     DivideByPowersOfTenWithPrecision(ascendTaskDatum.end) + DIVIDE_CHAR};
         // The original code contains a task start time filtering logic, of which is done in processor.
